changed logout to menulay component, added useref to mute
