--- src/components/logout/stylesLogout.css
@@ -1,32 +1,0 @@
-.button-logout{
-    display: flex;
-    width: 100%;
-    justify-content: end;
-    position: absolute;
-    z-index: 1;
-}
-.button-logout button {
-    cursor: pointer;
-}
-
-.button-next-lvl {
-    position: absolute;
-    bottom: 10px; /* Ajusta la distancia del borde inferior */
-    right: 10px; /* Ajusta la distancia del borde derecho */
-    z-index: 1;
-}
-
-.button-next-lvl button {
-    cursor: pointer;
-}
-
-.button-last-lvl {
-    position: absolute;
-    bottom: 10px; /* Ajusta la distancia del borde inferior */
-    left: 10px; /* Ajusta la distancia del borde izquierdo */
-    z-index: 1;
-}
-
-.button-last-lvl button {
-    cursor: pointer;
-}
